Move Telemetry classes to distinct package
