+++ src/main/java/org/jabref/gui/telemetry/Telemetry.java
@@ -1,12 +1,13 @@
-package org.jabref.gui;
+package org.jabref.gui.telemetry;
 
 import java.util.Optional;
 import java.util.TimerTask;
 
+import org.jabref.gui.DialogService;
+import org.jabref.gui.JabRefExecutorService;
 import org.jabref.gui.util.DefaultTaskExecutor;
 import org.jabref.logic.l10n.Localization;
 import org.jabref.logic.util.BuildInfo;
-import org.jabref.preferences.TelemetryPreferences;
 
 public class Telemetry {
     private Telemetry() {
@@ -24,7 +25,7 @@ public class Telemetry {
         });
     }
 
-    static void initTrackingNotification(DialogService dialogService, TelemetryPreferences preferences) {
+    public static void initTrackingNotification(DialogService dialogService, TelemetryPreferences preferences) {
         if (preferences.shouldAskToCollectTelemetry()) {
             JabRefExecutorService.INSTANCE.submit(new TimerTask() {
                 @Override
